Move status-writing responsibility out of helper function
In order to better conform with the changes made in #186096 (which was
created after but merged before this branch), we do not
write a rule status from within the check for unmatched index patterns,
but instead return a warning message that the caller then handles.

This allows the rule executor to decide what status should be written,
and without this change we were writing both a partial_failure (due to
the unmatched pattern), and also a success (due to the executor not
knowing about the warning).