[Infra] Fix dateRange handling when set to `now` (#206441)
## Summary

Quick fix to allow the datepicker on Infra pages to not break when
setting the `to` date to `"now"`. `use_date_picker()` hook already does
date parsing, so whatever it provides as the `dateRange` result is going
to be correct and not need further `Date` handling.

## How to test

* Go to Infra inventory page. Click on a host to open the flyout.
* Set the date for `to` to "now" on the Overview.
* It should update the page instead of crashing
* Any permutation of date ranges, from absolute, absolute + now, and
relative, should not crash the page and should load the correct data.
* Test against all tabs for the flyout/page to ensure no date
configuration causes crashes.