[APM] Attemp to fix flaky test in input_only_package.spec.ts (#206432)
## Summary

Fixes https://github.com/elastic/kibana/issues/201069

This PR addresses a flaky test issue in the APM package policy test
suite.

At times, the latency of a service with ingested Synthtrace data was
incorrect because one of the transactions had not been processed by the
time the test expectation was evaluated.

Since there’s no way to confirm when the transaction has been processed
and indexed, a retry strategy has been added. It will attempt the check
3 times, with a 100ms delay between each attempt, in an effort to
mitigate the flakiness of the test.

We also considered the possibility of not checking the latency
specifically and simply verifying that its value is > 0. However, by
keeping the exact value, we maintain the strict validation that ensures
the correctness of the calculations for the latency.

(cherry picked from commit fd2d1ba3dca7fc00c2b2331b4822ed156d1da7df)