Replace direct hex color usage in Exception Item Conditions component(s)
This replaces the use hex colors (to style a border), swapping instead
for the now-preferred euiTheme context, as well as using `@emotion` when possible.

These changes are applied here to both the in-plugin and packaged
versions of this component, which are used in two different sections. It
also attempts to bring them more inline with one another by applying a
missing `eui-xScroll` class that did not exist on the plugin variant of
this component.