[Streams] Migrate to storage adapter (#206438)
Migrates the `.kibana_streams` index to be managed by the Storage Index
Adapter.

Additionally, `streams_crud` was refactored into a client that can be
used in routes (and at some point, other plugins).

Functional changes:

- The defaults for the types were removed. Setting defaults means that
we have to use `.parse()` everywhere to make sure we get the default
values. Without using parse, the current type guards make it seem like
the defaults have been added to a value.
- The storage adapter was simplified (no more enums, required or
multi-value until we need it)
- Write access is verified before deleting a stream
- We throw an error if an update from ingest -> wired or vice versa is
attempted

Notes:
- I changed `readStream` to `getStream` to avoid confusion around
`ReadStreamDefinition` and `readStream`. This was before I realized that
ReadStreamDefinition refers to the value returned from the `GET
/api/streams/{id}` endpoint. I've left this change in place for now as I
would consider `get` to be more commonly used verbiage (eg.
`esClient.get`), but happy to revert.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Elastic Machine <elasticmachine@users.noreply.github.com>
Co-authored-by: Joe Reuter <johannes.reuter@elastic.co>