[Cases][Case Observables] move route key down in the component tree (#206449)
Fixes https://github.com/elastic/kibana/issues/206275

There was an issue with Case Page header not updating correctly when
navigating directly from one case to another (through similar cases
feature).

The solution for that was to add a key to the Route component for the
case view, but it is suboptimal - all the components on that page would
end up being re-rendered.

Thats why this PR moves the key to the header itself, with the same
result effectively but without re-rendering the entire page.

(cherry picked from commit 0c4c8aa1634164234e6ba286b5e9cedd274065d6)