[Streams 🌊] Add processors validation and simulation gate (#206566)
## 📓 Summary

Closes https://github.com/elastic/streams-program/issues/66

This work adds changes to prevent invalid processors from being
submitted.
The main rule is that a simulation is performed before any add/edit
submission to guarantee that the processor config is valid.

This work also updates the simulation API to detect whether there is a
non-additive change in any simulated document.

@patpscal error reporting UI for add/edit is different since the
simulator is not visible for edit, I used a callout but we can easily
update this once there is a final design in place.

### Form validation + simulation

https://github.com/user-attachments/assets/f7fc351b-6efc-4500-8490-b7f1c85139bf

### Non-additive processors

https://github.com/user-attachments/assets/47b5b739-c2cf-4a74-93a8-6ef43521c7d4
(cherry picked from commit 6429c53597b203d0f4435909333131b00c01a996)