[APM] Set metric trend to an empty array if undefined in MobileLocationStats (#206762)
## Summary

Closes https://github.com/elastic/kibana/issues/206712

This PR addresses an issue with the `MobileLocationStats` component in
the Service Overview.

In one of our charts, we were passing data to a `Metric` component where
the `trend` property was set to `undefined`. By definition, it should be
an empty array if there's no value. Initially, this wasn’t an issue
because no error appeared visually. However, after an update to the
`@elastic/charts` library, a full-screen error was triggered.

Additionally, a conversation has been initiated with the maintainers of
`@elastic/charts` to explore making the typings stricter, as we didn’t
receive any TypeScript warnings, even though the `trend` property should
not be `undefined`.

|Before|After|
|-|-|
|![Screenshot 2025-01-15 at 13 56
56](https://github.com/user-attachments/assets/19e2ae8e-73ca-4c65-ae6f-1d4d5bce5c1e)|![Screenshot
2025-01-15 at 13 57
44](https://github.com/user-attachments/assets/89ee9eff-17d2-47f7-b38b-c95423e76d78)|

(cherry picked from commit 0dfe806aaf6bce4dd9050e70224496f6d4c03fca)