[product doc] implement `highlight` summarizer (#206578)
## Summary

Fix https://github.com/elastic/kibana/issues/205921

- Implements a new summary strategy for the product documentation, based
on `semantic_text` highlights
- set that new strategy as the default one

### Why ?

Until now, in case of excessive token count, we were using a LLM based
summarizer. Realistically, highlights will always be worse than calling
a LLM for a "in context summary", but from my testing, highlights seem
"good enough", and the speed difference (instant for highlights vs
multiple seconds, up to a dozen, for the LLM summary) is very
significant, and seems overall worth it.

The main upside with that change, given that requesting the product doc
will be waaaay faster, is that we can then tweak the assistant's
instruction to more aggressively call the product_doc tool between each
user message without the risk of the user experience being impacted
(waiting way longer between messages). - *which will be done as a
follow-up*

### How to test ?

Install the product doc, ask questions to the assistant, check the tool
calls (sorry, don't have a better option atm...)

Note: that works with both versions of the product doc artifacts, so
don't need the dev repository