🌊 Streams: Fix multi level implicit stream creation (#206766)
As discussed in
https://github.com/elastic/streams-program/issues/26#issuecomment-2592146590,
it's currently not possible to create a wired stream multiple levels
deep in the hierarchy with a single request, as its implicit parents
won't be created properly.

This PR is fixing this issue by recursively calling `upsertStream` for
the parent as long as necessary.

It also adds a validation for children specified in the routing to make
sure they don't skip levels.

(cherry picked from commit 352dc8f1d51d8a591bf4211b0948b1e7d884ede3)