Remove usage of transparentize()
We only use this in our Duration components, as they are custom. It's
worth noting that we've re-implemented this in three different places,
as you can see here, but we can deduplicate later. These are EuiSelects
where Seconds/Minutes/Hours are chosen for either rule scheduling or
suppression duration.

Instead of using this custom color, it's been replaced with
`backgroundBasePrimary`, making it more in light with other similar form
controls.

I also replaced some usages of styled-components, here.