Remove custom styling of labelAppend on ScheduleItem
In testing this CSS (to determine a replacement), I discovered that it's
adding a fixed margin-left to labelAppend, which may extend outside the
bounds of the form row. The margin size itself seemed arbitrary. The
container for both labels is also a fixed size.

Instead, I opted to change the containing flexbox to justify as
`spaceBetween`, maximizing the distance between the items. If the
incoming `labelAppend` prop needs to add additional styling, the caller
can add that (although in testing I did not see usage of this prop).