Remove list-header styles file
This contained now-outdated references to EuiThemeVars that needed to be
inlined to the consuming component (in order to take advantage of the
euiTheme context/hook).

On the ListHeader component, I was able to remove a negative margin by
using `bottomBorder="extended"` as per the documentation. However, due
to the way the content is currently constructed, we still need a
negative margin-top in order to offset a hardcoded EuiSpacer in the
EuiPageHeader component (for which a comment was included on the style).

I was not able to use the full `useEuiFontSize` helper as the
line-heights have changed significantly from the original layout, hence
the `font-size: ().fontSize` pattern.