[EDR Workflows] Workflow Insights - migrate to Signature field (#205323)
This PR adds checks to verify whether the signer_id is present in file
events stored in the ES, which serve as the foundation for generating
endpoint insights. Previously, we relied solely on the executable path,
which caused issues when a single AV generated multiple paths.

With these changes:
* If the `signer_id` exists in the file event, it will be used for
generating insights alongside the path
* For cases where the `signer_id` is unavailable (e.g., Linux, which
lacks signers), the executable path will still be used as an only value.




https://github.com/user-attachments/assets/8965efef-e962-485a-b20f-d2730cffcf10

---------

Co-authored-by: Joey F. Poon <joey.poon@elastic.co>