[Search] [Playground] fix semantic_text issue (#207054)
## Summary

This fixes an issue in playground where the generated query is using a
multi_match. This is because the field is now defined as a text field
and Playground was treating the field as a text field and using it in a
multi-match.

This fix detects if the field is declared in the mappings API as
semantic_text and what the model_id is.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit 7987527d31ea25169fd1167c3c7d7132fb3c83f5)