[8.x] [Search] [Playground] fix semantic_text issue (#207054) (#207073)
# Backport

This will backport the following commits from `main` to `8.x`:
- [[Search] [Playground] fix semantic_text issue
(#207054)](https://github.com/elastic/kibana/pull/207054)

<!--- Backport version: 9.4.3 -->

### Questions ?
Please refer to the [Backport tool
documentation](https://github.com/sqren/backport)

<!--BACKPORT [{"author":{"name":"Joe
McElroy","email":"joseph.mcelroy@elastic.co"},"sourceCommit":{"committedDate":"2025-01-17T13:20:04Z","message":"[Search]
[Playground] fix semantic_text issue (#207054)\n\n## Summary\r\n\r\nThis
fixes an issue in playground where the generated query is using
a\r\nmulti_match. This is because the field is now defined as a text
field\r\nand Playground was treating the field as a text field and using
it in a\r\nmulti-match.\r\n\r\nThis fix detects if the field is declared
in the mappings API as\r\nsemantic_text and what the model_id
is.\r\n\r\n---------\r\n\r\nCo-authored-by: kibanamachine
<42973632+kibanamachine@users.noreply.github.com>","sha":"7987527d31ea25169fd1167c3c7d7132fb3c83f5","branchLabelMapping":{"^v9.0.0$":"main","^v8.18.0$":"8.x","^v(\\d+).(\\d+).\\d+$":"$1.$2"}},"sourcePullRequest":{"labels":["release_note:skip","v9.0.0","backport:prev-minor"],"title":"[Search]
[Playground] fix semantic_text
issue","number":207054,"url":"https://github.com/elastic/kibana/pull/207054","mergeCommit":{"message":"[Search]
[Playground] fix semantic_text issue (#207054)\n\n## Summary\r\n\r\nThis
fixes an issue in playground where the generated query is using
a\r\nmulti_match. This is because the field is now defined as a text
field\r\nand Playground was treating the field as a text field and using
it in a\r\nmulti-match.\r\n\r\nThis fix detects if the field is declared
in the mappings API as\r\nsemantic_text and what the model_id
is.\r\n\r\n---------\r\n\r\nCo-authored-by: kibanamachine
<42973632+kibanamachine@users.noreply.github.com>","sha":"7987527d31ea25169fd1167c3c7d7132fb3c83f5"}},"sourceBranch":"main","suggestedTargetBranches":[],"targetPullRequestStates":[{"branch":"main","label":"v9.0.0","branchLabelMappingKey":"^v9.0.0$","isSourceBranch":true,"state":"MERGED","url":"https://github.com/elastic/kibana/pull/207054","number":207054,"mergeCommit":{"message":"[Search]
[Playground] fix semantic_text issue (#207054)\n\n## Summary\r\n\r\nThis
fixes an issue in playground where the generated query is using
a\r\nmulti_match. This is because the field is now defined as a text
field\r\nand Playground was treating the field as a text field and using
it in a\r\nmulti-match.\r\n\r\nThis fix detects if the field is declared
in the mappings API as\r\nsemantic_text and what the model_id
is.\r\n\r\n---------\r\n\r\nCo-authored-by: kibanamachine
<42973632+kibanamachine@users.noreply.github.com>","sha":"7987527d31ea25169fd1167c3c7d7132fb3c83f5"}}]}]
BACKPORT-->

Co-authored-by: Joe McElroy <joseph.mcelroy@elastic.co>