Remove inference_id field if no inference endpoint is selected (#205660)
## Summary

Currently, the `semantic_text` field supports a default `inference_id`,
meaning users are not required to explicitly select an inference
endpoint during mapping. However, a bug has been identified: if the
`Select inference Id` popover is not opened, the `inference_id` field
property remains as an empty string. This causes Elasticsearch (ES) to
throw an error, as it requires a value to be present if the property is
defined.

To address this issue, the proposed solution is to remove the
`inference_id` property from the `semantic_text` field during field
mapping if its value is empty.

### Screen Recording


https://github.com/user-attachments/assets/e8d8d471-7ff2-493e-8872-e42838579d44

---------

Co-authored-by: Matthew Kime <matt@mattki.me>