[8.x] [Snapshot and Restore] Logsdb snapshot functional tests (#206318) (#207297)
# Backport

This will backport the following commits from `main` to `8.x`:
- [[Snapshot and Restore] Logsdb snapshot functional tests
(#206318)](https://github.com/elastic/kibana/pull/206318)

<!--- Backport version: 9.4.3 -->

### Questions ?
Please refer to the [Backport tool
documentation](https://github.com/sqren/backport)

<!--BACKPORT [{"author":{"name":"Sonia Sanz
Vivas","email":"sonia.sanzvivas@elastic.co"},"sourceCommit":{"committedDate":"2025-01-21T07:02:52Z","message":"[Snapshot
and Restore] Logsdb snapshot functional tests (#206318)\n\nPart of
https://github.com/elastic/kibana/issues/203716\n\n## Summary\nThis PR
creates the tests for Logsdb in the Snapshot and Restore plugin\n\n*
Verify that users can create snapshots from an LogsDB index\n-
Explanation: It should be possible to create a snapshot of a Index\nwith
LogsDb mode from a regular repository. This test creates a\nrepository
and a index with LogsDb mode, creates a policy, runs the\npolicy and
verifies that the state of the snapshot is `Complete` and it\ncontains
the LogsDb index.\n\n* Verify that users can restore a LogsDB
snapshot.\n- Explanation: It should be possible to restore a snapshot of
a Index\nwith LogsDb mode from a regular repository. This test takes the
snapshot\ncreated in the previous test and restore it. It verifies that
it has\nbeen restored and the status is `Complete`.\n \n* Verify that
users can NOT create a source-only snapshot from a LogsDB\nindex
[Snapshot result would be \"Partial\"].\n- Explanation: ES doesn't allow
to create a snapshot in a source-only\nrepository for index with a
synthetic source. Under the hood LogsDb uses\nsynthetic source (there is
no `_source`). So, is expected that, when\ncreating a snapshot that
includes a LogsDb index the result would be\npartial since it won't be
able to create the snapshot of the LogsDb\nindex. To test that, the test
creates a source-only repository and a\nindex with LogsDb mode, creates
a policy, runs the policy and verifies\nthat the state of the snapshot
is `Partial`.\n\n* Verify that users can NOT restore a source-only
snapshot from a LogsDB\nindex.\n- Explanation: Since the running the
policy in the previous test hasn't\ncreate the snapshot for the LogsDb
index, the snapshot for that index\ncouldn't be restored. To verify
that, the test tries to restore the\nsnapshot from the previous step and
wait to have the following error\n`index [sourceonly-logsdb-index]
wasn't fully snapshotted -
cannot\nrestore`\n\n---------\n\nCo-authored-by: Matthew Kime
<matt@mattki.me>","sha":"d6522194b87e2f078786744d8a1b93663d7ddb63","branchLabelMapping":{"^v9.0.0$":"main","^v8.18.0$":"8.x","^v(\\d+).(\\d+).\\d+$":"$1.$2"}},"sourcePullRequest":{"labels":["Team:Kibana
Management","release_note:skip","Feature:Snapshot and
Restore","v9.0.0","backport:prev-minor"],"title":"[Snapshot and Restore]
Logsdb snapshot functional
tests","number":206318,"url":"https://github.com/elastic/kibana/pull/206318","mergeCommit":{"message":"[Snapshot
and Restore] Logsdb snapshot functional tests (#206318)\n\nPart of
https://github.com/elastic/kibana/issues/203716\n\n## Summary\nThis PR
creates the tests for Logsdb in the Snapshot and Restore plugin\n\n*
Verify that users can create snapshots from an LogsDB index\n-
Explanation: It should be possible to create a snapshot of a Index\nwith
LogsDb mode from a regular repository. This test creates a\nrepository
and a index with LogsDb mode, creates a policy, runs the\npolicy and
verifies that the state of the snapshot is `Complete` and it\ncontains
the LogsDb index.\n\n* Verify that users can restore a LogsDB
snapshot.\n- Explanation: It should be possible to restore a snapshot of
a Index\nwith LogsDb mode from a regular repository. This test takes the
snapshot\ncreated in the previous test and restore it. It verifies that
it has\nbeen restored and the status is `Complete`.\n \n* Verify that
users can NOT create a source-only snapshot from a LogsDB\nindex
[Snapshot result would be \"Partial\"].\n- Explanation: ES doesn't allow
to create a snapshot in a source-only\nrepository for index with a
synthetic source. Under the hood LogsDb uses\nsynthetic source (there is
no `_source`). So, is expected that, when\ncreating a snapshot that
includes a LogsDb index the result would be\npartial since it won't be
able to create the snapshot of the LogsDb\nindex. To test that, the test
creates a source-only repository and a\nindex with LogsDb mode, creates
a policy, runs the policy and verifies\nthat the state of the snapshot
is `Partial`.\n\n* Verify that users can NOT restore a source-only
snapshot from a LogsDB\nindex.\n- Explanation: Since the running the
policy in the previous test hasn't\ncreate the snapshot for the LogsDb
index, the snapshot for that index\ncouldn't be restored. To verify
that, the test tries to restore the\nsnapshot from the previous step and
wait to have the following error\n`index [sourceonly-logsdb-index]
wasn't fully snapshotted -
cannot\nrestore`\n\n---------\n\nCo-authored-by: Matthew Kime
<matt@mattki.me>","sha":"d6522194b87e2f078786744d8a1b93663d7ddb63"}},"sourceBranch":"main","suggestedTargetBranches":[],"targetPullRequestStates":[{"branch":"main","label":"v9.0.0","branchLabelMappingKey":"^v9.0.0$","isSourceBranch":true,"state":"MERGED","url":"https://github.com/elastic/kibana/pull/206318","number":206318,"mergeCommit":{"message":"[Snapshot
and Restore] Logsdb snapshot functional tests (#206318)\n\nPart of
https://github.com/elastic/kibana/issues/203716\n\n## Summary\nThis PR
creates the tests for Logsdb in the Snapshot and Restore plugin\n\n*
Verify that users can create snapshots from an LogsDB index\n-
Explanation: It should be possible to create a snapshot of a Index\nwith
LogsDb mode from a regular repository. This test creates a\nrepository
and a index with LogsDb mode, creates a policy, runs the\npolicy and
verifies that the state of the snapshot is `Complete` and it\ncontains
the LogsDb index.\n\n* Verify that users can restore a LogsDB
snapshot.\n- Explanation: It should be possible to restore a snapshot of
a Index\nwith LogsDb mode from a regular repository. This test takes the
snapshot\ncreated in the previous test and restore it. It verifies that
it has\nbeen restored and the status is `Complete`.\n \n* Verify that
users can NOT create a source-only snapshot from a LogsDB\nindex
[Snapshot result would be \"Partial\"].\n- Explanation: ES doesn't allow
to create a snapshot in a source-only\nrepository for index with a
synthetic source. Under the hood LogsDb uses\nsynthetic source (there is
no `_source`). So, is expected that, when\ncreating a snapshot that
includes a LogsDb index the result would be\npartial since it won't be
able to create the snapshot of the LogsDb\nindex. To test that, the test
creates a source-only repository and a\nindex with LogsDb mode, creates
a policy, runs the policy and verifies\nthat the state of the snapshot
is `Partial`.\n\n* Verify that users can NOT restore a source-only
snapshot from a LogsDB\nindex.\n- Explanation: Since the running the
policy in the previous test hasn't\ncreate the snapshot for the LogsDb
index, the snapshot for that index\ncouldn't be restored. To verify
that, the test tries to restore the\nsnapshot from the previous step and
wait to have the following error\n`index [sourceonly-logsdb-index]
wasn't fully snapshotted -
cannot\nrestore`\n\n---------\n\nCo-authored-by: Matthew Kime
<matt@mattki.me>","sha":"d6522194b87e2f078786744d8a1b93663d7ddb63"}}]}]
BACKPORT-->

Co-authored-by: Sonia Sanz Vivas <sonia.sanzvivas@elastic.co>