[Ingest pipelines] Set outsideClickCloses to false if processor is modified (#207173)
Closes https://github.com/elastic/kibana/issues/198469
## Summary

When a user is working on adding a processor and click outside the side
flyout before saving the changes, their work gets lost. This is
especially annoying if the user is working on a complex processor and
clicks out of the window unintentionally.

To avoid this, we can take advance of the `outsideClickCloses` prop.
When the form has been modified (is dirty), clicking outside the flyout
won't close the flyout, so the work done in the processor won't be lost.
If nothing has been changed `outsideClickCloses` is set to true.

https://github.com/user-attachments/assets/0d70e16d-d731-4b01-b39e-3026a9c89002

This solution has been verified with @jovana-andjelkovic.

(cherry picked from commit 8dd8acc1b64e39a5cc29b0cced48d475725217b0)