[Streams 🌊] Update simulator to assert fields & integration testing (#206950)
## 📓 Summary

Closes https://github.com/elastic/streams-program/issues/68

This work updates the way a simulation for processing is performed,
working against the `_ingest/_simulate` API.
This gives less specific feedback on the simulation failure (which
processor failed), but allows for a much more realistic simulation
against the index configuration.

This work also adds integration testing for this API.

## 📔 Reviewer notes

The API is poorly typed due to missing typing in the elasticsearch-js
library. #204175 updates the library with those typings, as soon as it's
merged I'll update the API.

## 🎥  Recordings

https://github.com/user-attachments/assets/36ce0d3c-b7de-44d2-bdc2-84ff67fb4b25
(cherry picked from commit 39bf5e646fcaf31702dfe9fb17942d5aaea528ab)