[Rule Migration] Add PackageClient to fetch integrations (#207197)
## Summary

Awaiting the changes to the EPR API to include datasets we temporarily
included a JSON file to simulate the expected outcome so we could move
forward while waiting.

Since the changes is now done, the calls to
https://epr.elastic.co/search? now returns data_streams in their
responses so this PR replaces the temp JSON with the proper
PackageClient implementation.

Tested a few migrations with both prebuilt rule matches and without, and
seems to be working as intended. The integration count also seems to be
consistent with what to be expected.