🌊 Streams: Speed up sampling (#207334)
The sampling API can easily get slow because it matches on runtime
fields.

This PR picks the low hanging fruits in speeding it up:
* Not tracking the total hits because that would require to scan all
documents (by default it goes up to 10k)
* Stopping the search after as many matches as requested were found

This is not sufficient as search can still be very slow if condition
matches are rare, as millions of docs might have to be scanned. This can
be improved further in follow-up PRs.

(cherry picked from commit 49b28eda6521b8dba75931f0cc9b595be1d017aa)